Vitest batch 37: test-coverage-stragglers (DOM-split)
Ports the stub-based coverage probes to Vitest (7 asserts across 6
sections): extractDocx routing, the oura-auth json-catch arrow,
export.js reader.onerror, the AbortSignal.any polyfill branch, the IDB
onerror rails sweep (blob-storage / cashu-wallet / wearables-store /
backup), and dna.js worker.onerror. The fs-backed fetch shim lets §10's
parseDNAFile reach worker creation before the stubbed Worker errors.

Four sections need a browser runtime and split into
test-coverage-stragglers-dom.js on the puppeteer runner:
  - §1 img.onerror     — needs a real `new Image()` decoder
  - §4 showConfirmDialog — needs a real DOM overlay + animationend
  - §7 handleSSELine   — the `data:`-prefixed SSE fixture only
    accumulates through the OpenAI-compatible streaming path; relies on
    provider/config state the Vitest runner doesn't carry
  - §9 cashu _openDB onerror — needs a genuinely fresh module load so
    `_db` is null (Vitest ignores the `?bust=` query, and §8 opens the
    cashu DB, leaving `_db` cached)

Vitest: 85 → 86 tests. Puppeteer originals: 29 → 28.
